\name{hmapextension-package}
\alias{hmapextension-package}
\alias{hmapextension}
\docType{package}
\title{
\packageTitle{hmapextension}
}
\description{
\packageDescription{hmapextension}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hmapextension}
\packageIndices{hmapextension}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hmapextension}

Maintainer: \packageMaintainer{hmapextension}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
