% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_hmap_detailed.R
\name{pl_hmap_detailed}
\alias{pl_hmap_detailed}
\title{Printable heatmap grob}
\usage{
pl_hmap_detailed(matf1, matf2, gmatf, clusterby = 1, sample_labels = NULL,
  sample_labels2 = NULL, feature_labels = NULL, gene_labels = NULL,
  product_labels = NULL, col_ord, row_ord)
}

