% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_hmap_detailed.R
\name{sh_hmap_detailed}
\alias{sh_hmap_detailed}
\title{Produces detailed heatmap}
\usage{
sh_hmap_detailed(rw, sample_labels = NULL, sample_labels2 = NULL,
  feature_labels = NULL, prefix = "")
}
\arguments{
\item{rw}{List of dataframes 
Takes a read.grouped.table() as input or a list of four dataframes (more data frames are ok but it only uses these):\n
Counts - Genewise counts of expression\n
Tail - Mean tail length\n
Tail_counts - Number of poly-A tails counted\n
Annotation - Information regarding the annotation information\n
     Gene name, chromosome, gene product, biotype etc...}

\item{sample_labels}{Sample labels}

\item{sample_labels2}{Sample labels (second plot)}

\item{feature_labels}{Feature labels}

\item{prefix}{Prefix for plot}
}
\value{
Returns a composable shiny app object
}
\details{
Workhorse function for this package.
}

